feat(phase-2): split dashboard into legacy/v2 with feature-flag routing
